SE_TRANSIENT_FIELD_NOT_RESTORED
Bug: The field net.madz.download.engine.DownloadProcess.metadataFile is
transient but isn't set by deserialization
This class contains a field that is updated at multiple places in the
class, thus it seems to be part of the state of the class. However,
since the field is marked as transient and not set in readObject or
readResolve, it will contain the default value in any deserialized
instance of the class.

Confidence: High, Rank: Troubling (14)
Pattern: SE_TRANSIENT_FIELD_NOT_RESTORED
Type: Se, Category: BAD_PRACTICE (Bad practice)