WA_NOT_IN_LOOP
Bug: Wait not in loop in
net.madz.download.engine.StateMachineTest.testTransition(Dumper,
DownloadProcess, StateMachineMetaData)
This method contains a call to java.lang.Object.wait() which is not in
a loop.  If the monitor is used for multiple conditions, the condition
the caller intended to wait for might not be the one that actually
occurred.

Confidence: Normal, Rank: Troubling (14)
Pattern: WA_NOT_IN_LOOP
Type: Wa, Category: MT_CORRECTNESS (Multithreaded correctness)